Only attempt to delete output if present

+++ index.js
@@ -3,59 +3,62 @@
  * from commans line to run Puppeteer web scrape scripts.
  */
 
-const yargs = require('yargs');
-const fs = require('fs');
+const yargs = require("yargs");
+const fs = require("fs");
 const csvOneDimArray = require("./csv-onedim-array"); // Loads CSV input and translates to array, element per row
 const downloadData = require("axios-https-snip"); // download from HTTP source
 const writeFileAsync = require("./writeFileAsync"); // write file locally
 const convertSitemap = require("sitemap-url-array"); // Converts XML sitemap for JSON input
 const listCrawler = require("./scripts/listCrawler"); // Puppeteer script to scrape links from a paginated listing page, to be used as input
 const scrollCrawler = require("./scripts/scrollCrawler"); // Puppeteer script to scrape links from a lazy loading listing page, to be used as input
-const _progress = require('cli-progress'); // Progress bar that will appear in STDOUT when scraper is running
+const _progress = require("cli-progress"); // Progress bar that will appear in STDOUT when scraper is running
 
 const main = async () => {
-// Load config
-const paramConfig = process.argv[2];
-const configFile = require('./config.json');
-let config = configFile.configs.find(e => e.configName == paramConfig);
-const basilScript = require(`./scripts/${config.script.name}`);
+  // Load config
+  const paramConfig = process.argv[2];
+  const configFile = require("./config.json");
+  let config = configFile.configs.find((e) => e.configName == paramConfig);
+  const basilScript = require(`./scripts/${config.script.name}`);
 
-// Get overrides from command line
-const argv = yargs
-    .option('parallel', {
-        describe: 'Number of parallel instances',
-        type: 'number',
+  // Get overrides from command line
+  const argv = yargs
+    .option("parallel", {
+      describe: "Number of parallel instances",
+      type: "number",
     })
-    .option('input', {
-        describe: 'File path to JSON list of URLs to scrape',
-        type: 'string',
+    .option("input", {
+      describe: "File path to JSON list of URLs to scrape",
+      type: "string",
     })
-    .option('urlSitemap', {
-        describe: 'URL of sitemap to use as input ',
-        type: 'string',
-    })
-    .argv;
+    .option("urlSitemap", {
+      describe: "URL of sitemap to use as input ",
+      type: "string",
+    }).argv;
 
-// Set properties, preferring command line params over config file where supplied.
-config.parallel = argv.parallel || config.parallel;
-config.input = argv.input || config.input;
-config.urlSitemap = argv.urlSitemap || config.urlSitemap;
-console.log(config);
+  // Set properties, preferring command line params over config file where supplied.
+  config.parallel = argv.parallel || config.parallel;
+  config.input = argv.input || config.input;
+  config.urlSitemap = argv.urlSitemap || config.urlSitemap;
+  console.log(config);
 
-/* Set up input source for running scripts. 
+  /* Set up input source for running scripts. 
 Pass an array of URLs to the scrpt module.
 */
 
-const filePath = "./input/sitemap.xml"; // Path to store sitemap XML
-const jsonSitemap = "./input/sitemap.json"; // Path to store sitemap coverted to JSON
-const outPath = typeof config.outputPath == "undefined" ? "./output/webscrape.csv" : config.outputPath; // Optional output file location
-const followRedirect = typeof config.followRedirect == "undefined" ? true : config.followRedirect; // Option to follow redirects when scraping
+  const filePath = "./input/sitemap.xml"; // Path to store sitemap XML
+  const jsonSitemap = "./input/sitemap.json"; // Path to store sitemap coverted to JSON
+  const outPath =
+    typeof config.outputPath == "undefined"
+      ? "./output/webscrape.csv"
+      : config.outputPath; // Optional output file location
+  const followRedirect =
+    typeof config.followRedirect == "undefined" ? true : config.followRedirect; // Option to follow redirects when scraping
 
-/* Get input of URLs for both input path, sitemap, and scrape of a listing page, depending on what config specifies.
+  /* Get input of URLs for both input path, sitemap, and scrape of a listing page, depending on what config specifies.
   Combine them into a single input.
   */
   const arrPages = [];
-  const {input, urlSitemap, pageList, scrollList} = config;
+  const { input, urlSitemap, pageList, scrollList } = config;
 
   // Get URLs specified by input path
   if (input) {
@@ -89,10 +92,10 @@ const followRedirect = typeof config.followRedirect == "undefined" ? true : conf
       arrPages.push(...arrListLinks);
     } catch (error) {
       console.error("Error:", error.message || error);
-    };
-  };
+    }
+  }
 
-   /* Get URLs from scraping a list on the target website, such as a product listing.
+  /* Get URLs from scraping a list on the target website, such as a product listing.
   List type: lazy loading scroll
   */
 
@@ -102,24 +105,30 @@ const followRedirect = typeof config.followRedirect == "undefined" ? true : conf
       arrPages.push(...arrListLinks);
     } catch (error) {
       console.error("Error:", error.message || error);
-    };
-  };
+    }
+  }
 
   const arrUniquePages = [...new Set(arrPages)]; // Remove duplicate from array of URLs
   config.arrUniquePages = arrUniquePages;
 
   // Remove existing output file if present
-  fs.unlink(outPath, (err) => {
-    if (err) {
-      console.error(err);
-    };
-  });
+  if (fs.existsSync(outPath)) {
+    fs.unlink(outPath, (err) => {
+      if (err) {
+        console.error(err);
+      }
+    });
+  };
 
   // create a new progress bar to pass to scraper
-  config.bar = new _progress.Bar({
-    format: 'progress [{bar}] {percentage}% | ETA: {eta_formatted} | Duration: {duration_formatted} | {value}/{total}'
-  }, _progress.Presets['shades_classic']);
-  
+  config.bar = new _progress.Bar(
+    {
+      format:
+        "progress [{bar}] {percentage}% | ETA: {eta_formatted} | Duration: {duration_formatted} | {value}/{total}",
+    },
+    _progress.Presets["shades_classic"]
+  );
+
   return basilScript(config);
 };
